refactor(cubeapi): move exec-code endpoint from e2b surface to /cube/ prefix
