+++ CubeAPI/src/handlers/mod.rs
@@ -6,6 +6,7 @@ pub mod agenthub;
 pub mod auth;
 pub mod cluster;
 pub mod config;
+pub mod cube;
 pub mod examples;
 pub mod health;
 pub mod sandboxes;
